refactor: isolate web update internals
Move the web updater state machine out of entrypoints and routes so deeper cleanup stays localized without changing update behavior. Keep the same web update APIs while reducing duplication across UI and server flows.
